+++ tests/Moment/MustBeAfterOrEqualToTest.php
@@ -1,0 +1,74 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCraftsman\DateTimePrecision\Moment;
+
+use DigitalCraftsman\DateTimePrecision\Exception\MomentIsNotAfterOrEqualTo;
+use DigitalCraftsman\DateTimePrecision\Moment;
+use DigitalCraftsman\DateTimePrecision\Test\Exception\CustomMomentIsNotAfterOrEqualTo;
+use PHPUnit\Framework\TestCase;
+
+/** @coversDefaultClass \DigitalCraftsman\DateTimePrecision\Moment */
+final class MustBeAfterOrEqualToTest extends TestCase
+{
+    /**
+     * @test
+     *
+     * @param ?class-string<\Throwable> $expectedResult
+     *
+     * @dataProvider dataProvider
+     *
+     * @covers ::mustBeAfterOrEqualTo
+     */
+    public function must_be_after_or_equal_to_works(
+        ?string $expectedResult,
+        Moment $moment,
+        Moment $comparator,
+        ?callable $otherwiseThrow,
+    ): void {
+        // -- Act & Assert
+        if ($expectedResult !== null) {
+            $this->expectException($expectedResult);
+        } else {
+            $this->expectNotToPerformAssertions();
+        }
+
+        $moment->mustBeAfterOrEqualTo(
+            $comparator,
+            $otherwiseThrow,
+        );
+    }
+
+    /**
+     * @return array<string, array{
+     *   0: ?string,
+     *   1: Moment,
+     *   2: Moment,
+     *   3: ?callable(): \Throwable
+     * }>
+     */
+    public static function dataProvider(): array
+    {
+        return [
+            'without exception' => [
+                null,
+                Moment::fromString('2022-10-08 16:00:00'),
+                Moment::fromString('2022-10-08 15:00:00'),
+                null,
+            ],
+            'default exception' => [
+                MomentIsNotAfterOrEqualTo::class,
+                Moment::fromString('2022-10-08 14:00:00'),
+                Moment::fromString('2023-10-08 15:00:00'),
+                null,
+            ],
+            'custom exception' => [
+                CustomMomentIsNotAfterOrEqualTo::class,
+                Moment::fromString('2022-10-08 14:00:00'),
+                Moment::fromString('2022-10-08 15:00:00'),
+                static fn () => new CustomMomentIsNotAfterOrEqualTo(),
+            ],
+        ];
+    }
+}